Simple GET request for items.  Not integrated w a database yet
